+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Accessibility 77 Further changes to prevent horizontal scrolling on smaller mobile devices
+- Add social media text & new icon/links to footer
 
 ## 22.0 (2020-07-23)